Исправлен код

+++ models/metric/models.py
@@ -8,7 +8,7 @@ class ServerStat(models.Model):
     # server = models.IntegerField(null=True, blank=True)
    
     Use_Cpu = models.IntegerField(help_text="CPU(%)")#,null=True, blank=True
-    Use_Ram = models.IntegerField(help_text="RAM(%)")
+    Use_Ram = models.FloatField(help_text="RAM(%)")
     Procent_Ram = models.FloatField(help_text="Procent RAM(%)",null=True, blank=True)
     Use_Swap = models.IntegerField(help_text="SWAP")
     # IO = models.FloatField(help_text="Input/Output(%)")
